Debugging sensors

+++ QController2/Orientation.pde
@@ -9,8 +9,9 @@ void updateOrientation(int deltaTime) {
     // multiply with elapsed time to get rotation since last update
     // 5000mV / 1024 (10-bit ADC resolution) = 4.8828125 mV for each ADC value
     // 5mv / degree / second
-    int degPerSec = ((GYRO_RAW[n] * 4.8828125) / 5);
+    float degPerSec = ((GYRO_RAW[n] * 4.8828125) / 5);
     GYRO_ANGLE[n] += ((degPerSec / 1000.0) * deltaTime);
+//    GYRO_ANGLE[n] += ((GYRO_RAW[n] / 1000.0) * deltaTime);
   }
   
   // Update bias every 20 ms
